requirements.txt
python-dotenv
Flask
gunicorn
requests
python-dotenv